RoleDiscord/requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
discord.py==1.4.1
dnspython==1.15.0
dnspython3==1.15.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mcstatus==4.1.0
multidict==4.7.6
pylint==2.5.3
six==1.15.0
toml==0.10.1
wrapt==1.12.1
yarl==1.5.1
